windows / non-windows tests
